% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{add_api_key_header}
\alias{add_api_key_header}
\title{Adds API key header to the qdrant connection for security}
\usage{
add_api_key_header(req, conn)
}
\arguments{
\item{req}{a httr2 request object that is prepared for Qdrant REST API calls.}

\item{conn}{a connection object created by get_qdrant_connection()}
}
\description{
\code{add_api_key_header} adds a security header to connect to Qdrant instance with an API key.
}
\details{
Adds a security header to connect to Qdrant instance with an API key.
The API key value is provided by the connection object obtained with get_qdrant_connection().
This function returns the request object along with the necessary header.
}
