% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{add_tools_declaration}
\alias{add_tools_declaration}
\title{Add tools declaration for the LLM to use}
\usage{
add_tools_declaration(workflow_obj, tools)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the workflow required}

\item{tools}{a list of tools declared a R list, see examples.}
}
\description{
\code{split_text_as_sentences} lets you add some tools (functions) that can be used if needed by the LLM to answer questions
}
\details{
Lets you add some tools (functions) that can be used by the LLM.
This only works for models where the tool calling function is supported, which is the case for Llama3.1 for example.
The structure expected is a R list.
}
\examples{
tool_list <- list(
 list(type="function",
      "function"=list(
        name="get_flight_times",
        description="get the flight times between two cities",
        parameters= list(
          type="object",
          properties = list(
            departure=list(
              type="string",
              description="the departure city (airport code)"
            ),
            arrival=list(
              type="string",
              description="the arrival city (airport code)"
            )
          ),
          required=c("departure", "arrival")
        )
      )))

myflow_test <- ai_workflow() |>
   set_connector("ollama")  |> 
   set_model(model_name= "llama3.1:8b-instruct-q5_K_M") |>
   set_n_predict(1000) |>
   set_temperature(0.8) |> 
   set_default_missing_parameters_in_workflow() |> 
   add_tools_declaration(tool_list)
   
}
