% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{add_workflow_step_old}
\alias{add_workflow_step_old}
\title{Add a step (i.e. another workflow) to an existing workflow}
\usage{
add_workflow_step_old(workflow_obj, workflow_obj_to_add, type = "chain")
}
\arguments{
\item{workflow_obj}{the previous workflow object that you want to build on}

\item{workflow_obj_to_add}{the workflow object you want to add on top of the existing one}

\item{type}{the type of step you want to add to the existing workflow. Defaults to "chain".}
}
\description{
\code{add_workflow_step_old} adds another workflow to an existing one. By default it chains the new workflow to the previous one(s).
}
\details{
This function will add a new workflow to an existing one. By default the way the new workflow is added is by chaining it to the previous one.
The way this works is that it will use the previous output of the last workflow element as input for the next one.
}
\examples{
myflow_template <- ai_workflow() |> 
set_connector("ollama") |>
  set_model(model_name= "llama3.1:8b-instruct-q5_K_M") |> 
  set_n_predict(1000) |> 
  set_temperature(0.8) 


}
