% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{ai_workflow}
\alias{ai_workflow}
\title{Define AI workflow}
\usage{
ai_workflow()
}
\value{
an empty list in the ai_workflow() format that will be used to capture all settings and parameters of the workflow.
}
\description{
\code{ai_workflow} creates a list that encapsulates parameters related to the AI workflow you want to design.
}
\details{
The starting point of defining an AI workflow in this package. A workflow object contains a connector and a model as placeholders.
}
\examples{
my_workflow <- ai_workflow()
}
