% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiworkflow-package.R
\docType{package}
\name{aiworkflow-package}
\alias{aiworkflow}
\alias{aiworkflow-package}
\title{aiworkflow: Enable R users to leverage AI frameworks like Ollama and Llama.cpp}
\description{
ai_workflow is a R package that makes it easy to run typical AI-powered (mostly LLM-related) tasks with R. Its main focus is on Local AI support, while it can be expanded to third-party LLMs in the future.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-guyot.github.io/aiworkflow/}
}

}
\author{
\strong{Maintainer}: Raphael Guyot \email{rguyot@sanqualis.com}

}
\keyword{internal}
