% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{apply_processing_skill}
\alias{apply_processing_skill}
\title{Applies a processing skill to the current workflow}
\usage{
apply_processing_skill(
  prompts_vector,
  processing_skill = NA,
  processing_skill_args = list()
)
}
\arguments{
\item{prompts_vector}{a vector containing existing prompts or text on which the skill will be applied.}

\item{processing_skill}{character vector containing the filepath to the processing skill text template of the skill to be applied.}

\item{processing_skill_args}{a list of additional parameters to feed to a processing skill, if it accepts them.}
}
\value{
A new vector containing the modified prompts after applying the processing skill template.
}
\description{
\code{apply_processing_skill} applies a processing skill (i.e. a pre-engineered prompt)
}
\details{
A simple function to apply a pre-defined prompt format (i.e. a processing skill) to the current workflow.
You can find out about the existing processing skills by using the list_processing_skills() function.
This function is not usually made to be used on its own, but through the set_processing_skill() function
}
