% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_check_queue_prompt_status}
\alias{cfy_check_queue_prompt_status}
\title{ComfyUI: Check the queued prompts statuses}
\usage{
cfy_check_queue_prompt_status(workflow_obj, prompt_id)
}
\description{
\code{cfy_check_queue_prompt_status} lets you check the status of the queued prompts from ComfyUI.
}
