% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_find_pos_neg_prompt_node}
\alias{cfy_find_pos_neg_prompt_node}
\title{ComfyUI: Find the Node of the Negative Prompt}
\usage{
cfy_find_pos_neg_prompt_node(workflow_obj, polarity)
}
\description{
\code{cfy_find_pos_neg_prompt_node} lets you find the node reference of the negative prompt from ComfyUI.
}
