% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_get_history}
\alias{cfy_get_history}
\title{ComfyUI: Get History}
\usage{
cfy_get_history(workflow_obj, prompt_id)
}
\description{
\code{cfy_get_history} lets you get the history of the images generated by the current ComfyUI workflow.
}
