% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_get_image}
\alias{cfy_get_image}
\title{ComfyUI: Get Generated Image}
\usage{
cfy_get_image(workflow_obj, image_filename)
}
\description{
\code{cfy_get_image} lets you get a generated image from ComfyUI.
}
