% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_get_model_checkpoints}
\alias{cfy_get_model_checkpoints}
\title{ComfyUI: Get Model Checkpoints}
\usage{
cfy_get_model_checkpoints(workflow_obj)
}
\description{
\code{cfy_get_model_checkpoints} lets you get a list of available checkpoint models from ComfyUI.
}
