% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_get_pictures_list}
\alias{cfy_get_pictures_list}
\title{ComfyUI: Get Pictures List}
\usage{
cfy_get_pictures_list(workflow_obj, prompt_id)
}
\description{
\code{cfy_get_pictures_list} lets you get a list pictures already generated from ComfyUI.
}
