% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_process_prompts}
\alias{cfy_process_prompts}
\title{ComfyUI: Process Prompts}
\usage{
cfy_process_prompts(workflow_obj, prompt)
}
\description{
\code{cfy_process_prompt} lets you process the prompts to generate the image with the current ComfyUI workflow.
}
