% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_queue_prompt}
\alias{cfy_queue_prompt}
\title{ComfyUI: Queue Prompt}
\usage{
cfy_queue_prompt(workflow_obj, prompt_json)
}
\description{
\code{cfy_queue_prompt} lets you queue the prompt to generate the image with the current ComfyUI workflow.
}
