% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_cfg}
\alias{cfy_set_cfg}
\title{ComfyUI: set cfg (guidance)}
\usage{
cfy_set_cfg(workflow_obj, cfg)
}
\description{
\code{cfy_set_cfg} lets you set the guidance (cfg) for the present comfyUI workflow
}
