% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_checkpoint}
\alias{cfy_set_checkpoint}
\title{ComfyUI: set Checkpoint}
\usage{
cfy_set_checkpoint(workflow_obj, checkpoint)
}
\description{
\code{cfy_set_checkpoint} lets you set the Checkpoint Model to use for the present comfyUI workflow
}
