% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_custom_workflow}
\alias{cfy_set_custom_workflow}
\title{ComfyUI: set custom workflow}
\usage{
cfy_set_custom_workflow(workflow_obj, comfyui_workflow_json_filepath)
}
\description{
\code{cfy_set_custom_workflow} lets you load a specific json file that contains a custom comfyui workflow
}
