% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_image_size}
\alias{cfy_set_image_size}
\title{ComfyUI: set image size}
\usage{
cfy_set_image_size(workflow_obj, resolution)
}
\description{
\code{cfy_set_image_size} lets you set the image size for the present comfyUI workflow
}
