% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_lora_clip_strength}
\alias{cfy_set_lora_clip_strength}
\title{ComfyUI: set Strength of the CLIP for LORA}
\usage{
cfy_set_lora_clip_strength(workflow_obj, strength_clip)
}
\description{
\code{cfy_set_lora_clip_strength} lets you set the strength of the CLIP Modification by the LORA for the present comfyUI workflow
}
