% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_lora_model_strength}
\alias{cfy_set_lora_model_strength}
\title{ComfyUI: set LORA Model Strength}
\usage{
cfy_set_lora_model_strength(workflow_obj, strength_model)
}
\description{
\code{cfy_set_lora_model_strength} lets you set the strength of the LORAL model for the present comfyUI workflow
}
