% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_negative_prompt}
\alias{cfy_set_negative_prompt}
\title{ComfyUI: Set Negative Prompt for Image Generation}
\usage{
cfy_set_negative_prompt(workflow_obj, negative_prompt)
}
\description{
\code{cfy_set_negative_prompt} lets you set the negative prompt to use to generate images in ComfyUI.
}
