% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_positive_prompt}
\alias{cfy_set_positive_prompt}
\title{ComfyUI: Set Positive Prompt for Image Generation}
\usage{
cfy_set_positive_prompt(workflow_obj, positive_prompt)
}
\description{
\code{cfy_set_positive_prompt} lets you set the positive prompt to use to generate images in ComfyUI.
}
