% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_sampler}
\alias{cfy_set_sampler}
\title{ComfyUI: set sampler}
\usage{
cfy_set_sampler(workflow_obj, sampler)
}
\description{
\code{cfy_set_sampler} lets you set the sampler for the present comfyUI workflow
}
