% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_scheduler}
\alias{cfy_set_scheduler}
\title{ComfyUI: set scheduler}
\usage{
cfy_set_scheduler(workflow_obj, scheduler)
}
\description{
\code{cfy_set_scheduler} lets you set the scheduler for the present comfyUI workflow
}
