% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_seed_change}
\alias{cfy_set_seed_change}
\title{ComfyUI: set Seed Change Parameter}
\usage{
cfy_set_seed_change(workflow_obj, seed_change)
}
\description{
\code{cfy_set_seed_change} lets you set the seed change parameter to use for the present comfyUI workflow
}
