% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_simple_worfklow}
\alias{cfy_set_simple_worfklow}
\title{ComfyUI: Set up a very simple ComfyUI workflow}
\usage{
cfy_set_simple_worfklow(
  workflow_obj,
  checkpoint = NA,
  steps = 20,
  seed = sample(1:1e+10, 1),
  scheduler = "normal",
  sampler = "euler"
)
}
\description{
\code{cfy_set_simple_workflow} lets you set most simple ComfyUI workflow possible
}
