% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_simple_workflow_negative_prompt}
\alias{cfy_set_simple_workflow_negative_prompt}
\title{ComfyUI: set a negative prompt for a simple ComfyUI workflow}
\usage{
cfy_set_simple_workflow_negative_prompt(workflow_obj, negative_prompt)
}
\description{
\code{cfy_set_simple_workflow_negative_prompt} lets you set the negative prompt for a simple ComfyUI workflow
}
