% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_set_steps}
\alias{cfy_set_steps}
\title{ComfyUI: set number of steps for image generation}
\usage{
cfy_set_steps(workflow_obj, steps = 20)
}
\description{
\code{cfy_set_steps} lets you set the number of steps to generate an image with the present comfyUI workflow
}
