% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfy.R
\name{cfy_unload_model}
\alias{cfy_unload_model}
\title{ComfyUI: Unload model}
\usage{
cfy_unload_model(workflow_obj)
}
\description{
\code{cfy_unload_model} lets you unload a model stored in VRAM to free memory.
}
