% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{change_all_seeds}
\alias{change_all_seeds}
\title{Change all seeds across a workflow at once}
\usage{
change_all_seeds(workflow_obj, fixed_seed = NA)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the flow required}

\item{fixed_seed}{A numerical value (defaults to NA, which means a random seed is generated) to fix the new seed.}
}
\description{
\code{change_all_seeds} is a way to change all seeds at once for a workflow
}
\details{
This function provides a way to change all seeds at once for a workflow, either randomly or with a fixed seed.
}
