% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{convert_embeddings_to_qdrant_format}
\alias{convert_embeddings_to_qdrant_format}
\title{Convert embeddings to Qdrant format}
\usage{
convert_embeddings_to_qdrant_format(input)
}
\arguments{
\item{input}{embeddings resulting from the generate_document_embeddings() function.}
}
\description{
\code{convert_embeddings_to_qdrant_format} converts embeddings to the expected qdrant format for upserting.
'
}
\details{
Embeddings generated by the generate_document_embeddings() function have a different structure.
To make such embeddings directly usable with the qdrant function we can use this to make it seamless.
}
\examples{
sentence <- "hi there this is a great day"
tmp_path <- tempfile()
writeLines(sentence, tmp_path)
conn <- get_ollama_connection()
doc_embeddings <- generate_document_embeddings(ollama_connection = conn,document_path = tmp_path)
embeddings_for_qdrant <- doc_embeddings |> convert_embeddings_to_qdrant_format()

}
