% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{convert_ollama_completion_response_to_tibble}
\alias{convert_ollama_completion_response_to_tibble}
\title{Convert an ollama server completion response to a tibble}
\usage{
convert_ollama_completion_response_to_tibble(ollama_response)
}
\arguments{
\item{ollama_response}{This should contain the full response of a ollama API completion call.}
}
\description{
\code{get_ollama_completion} get a completion from the ollama server API
}
\details{
A function to convert the full output of ollama to a tibble format
}
