% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{convert_ollama_model_info_response_to_tibble}
\alias{convert_ollama_model_info_response_to_tibble}
\title{Convert ollama response for model info to a tibble}
\usage{
convert_ollama_model_info_response_to_tibble(ollama_response)
}
\arguments{
\item{ollama_response}{a response object coming from a payload delivered by the ollama server instance.}
}
\description{
\code{convert_ollama_model_info_response_to_tibble} Converts the ollama response for model info to a tibble
}
\details{
Converts the ollama response for model info to a tibble
}
