% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{create_custom_processing_skill}
\alias{create_custom_processing_skill}
\title{Create a processing skill template file that can be used to create your own skills}
\usage{
create_custom_processing_skill(
  filepath,
  system_prompt = NA,
  chat_prompt = NA,
  final_guidance = NA
)
}
\arguments{
\item{filepath}{the file path where you'd like to save the processing skill. It needs to be a ".txt" file.}

\item{system_prompt}{the system prompt to give to the processing skill. Optional. Defaults to NA. You are advised to put few shots examples in there.}

\item{chat_prompt}{the chat prompt to give to the processing skill. Optional. Defaults to NA.}

\item{final_guidance}{the final recommendations you want to give to the processing skill at the end. Optional. Defaults to NA.}
}
\description{
\code{create_custom_processing_skill} creates a text file with a processing skill template.
}
\details{
This function creates a processing skill text file based on a template.
If you use the arguments for system_prompt, chat_prompt and final_guidance, they all need to be present at once. If not, it will default to the standard template.
}
\examples{
create_custom_processing_skill("write_poem.txt")
}
