% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{decapsulate}
\alias{decapsulate}
\title{Decapsulate removes the encapsulation of a workflow to make it easier to edit}
\usage{
decapsulate(workflow_obj)
}
\arguments{
\item{workflow_obj}{the workflow_obj that you want to encapsulate to be able to make use of it}
}
\description{
\code{decapsulate} removes the encapsulation of a workflow to make it easier to edit.
}
\details{
This function removes the encapsulation of a workflow to make it easier to edit.
}
\examples{
myflow_template <- ai_workflow() |> 
set_connector("ollama") |>
  set_model(model_name= "llama3.1:8b-instruct-q5_K_M") |> 
  set_n_predict(1000) |> 
  set_temperature(0.8) |>
  encapsulate()
  
myflow_template <- myflow_template |>
decapsulate() |>
add_processing_skill("add_details") |>
encapsulate()

}
