% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{display_intermediate_answer}
\alias{display_intermediate_answer}
\title{Display Intermediate Answer}
\usage{
display_intermediate_answer(workflow)
}
\arguments{
\item{workflow}{A workflow object containing all parameters describing the flow required}
}
\description{
\code{display_intermediate_answer} is a function that displays the results of the last executed workflow
}
\details{
This function will print the output of the last workflow executed. Especially helpful when chaining several workflows.
}
