% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{encapsulate}
\alias{encapsulate}
\title{Encapsulates a workflow for it to be used to process prompts}
\usage{
encapsulate(workflow_obj)
}
\arguments{
\item{workflow_obj}{the workflow_obj that you want to encapsulate to be able to make use of it}
}
\description{
\code{encapsulate} readies a workflow to process prompts.
}
\details{
This function encapsulates a workflow to get it ready to be used to process prompts.
}
\examples{
myflow_template <- ai_workflow() |> 
set_connector("ollama") |>
  set_model(model_name= "llama3.1:8b-instruct-q5_K_M") |> 
  set_n_predict(1000) |> 
  set_temperature(0.8) |>
  encapsulate()

}
