% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{extract_snippets}
\alias{extract_snippets}
\title{Extract Snippets}
\usage{
extract_snippets(text)
}
\arguments{
\item{text}{the text to extract snippets from.}
}
\description{
\code{extract_snippets} makes it easy to extract snippets in generated answers from the LLM
}
\details{
This function makes it easy to extract snippets in generated answers from the LLM
}
\examples{
llm_answer <- "This is a simple function in r:
```
yo <- function(message) {
print(message)
}
```
Enjoy this function!"

extracted_function <- llm_answer |> extract_snippets()
}
