% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{generate_document_embeddings}
\alias{generate_document_embeddings}
\title{Get embeddings for a piece of context through an ollama server instance020}
\usage{
generate_document_embeddings(
  ollama_connection,
  document_path,
  splitter = "sentence",
  model = "bge-large"
)
}
\arguments{
\item{ollama_connection}{a connection object that has the information on how to connect to the ollama server}

\item{document_path}{the document path to create embeddings from}

\item{splitter}{the type of splitter that is going to be used. Defaults to "sentence". "paragraph" can also be used.}

\item{model}{The model used to generates the embeddings. Defaults to "bge-large". You can select any other model available for generating embeddings.}
}
\description{
\code{generate_document_embeddings} Generates vector embeddings for a document through ollama
}
\details{
Converts the ollama response for model info to a tibble
}
\examples{
#conn<-get_ollama_connection()
#embeddings<- get_ollama_embeddings(ollama_connection=conn, prompt="Roma is a beautiful city.")
}
