% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{generate_numeric_list}
\alias{generate_numeric_list}
\title{Generate Numeric List}
\usage{
generate_numeric_list(decimals = 1, length)
}
\arguments{
\item{decimals}{the number of decimals needed for each figure. Defaults to 1.}

\item{length}{the length of the list you want to produce}
}
\description{
\code{generate_numeric_list} generates a random list to simulate what vector embeddings look like
}
\details{
Generates a list of numerical values of a specific length.
It can be used to simulate what vector embeddings look like.
}
\examples{
generate_numeric_list(decimals=1, length=384)

}
