% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{generate_uuid_from_text}
\alias{generate_uuid_from_text}
\title{Generate UUID from text}
\usage{
generate_uuid_from_text(text = NA_character_)
}
\arguments{
\item{text}{a piece of text to generate a UUID. Defaults to NA.}
}
\description{
\code{generate_uuid_from_text} generates a unique UUID for a given piece of text.
}
\details{
Generates a UUID for a piece of text. The same text will give the same id.
This is useful to generate an id for inserting vectors into qdrant.
}
\examples{
generate_uuid_from_text("hi there")

}
