% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{get_list_ollama_models}
\alias{get_list_ollama_models}
\title{Get a list of models available from the ollama server}
\usage{
get_list_ollama_models(ollama_connection)
}
\arguments{
\item{ollama_connection}{an ollama connection objection object that was created by get_ollama_connection().}
}
\value{
A dataframe that contains information about the LLM available from the running ollama server
}
\description{
\code{get_list_ollama_models} gets a list of models available from the ollama server
}
\details{
A simple function to get the models available from the ollama server.
}
\examples{
ollama_conn <- get_ollama_connection()
list_of_models <- get_list_ollama_models(ollama_conn)

}
