% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{get_ollama_connection}
\alias{get_ollama_connection}
\title{Define a connection to a local ollama server}
\usage{
get_ollama_connection(ip_ad = "127.0.0.1", port = "11434")
}
\arguments{
\item{ip_ad}{the IP address of the server running ollama. Default is localhost 127.0.0.1.}

\item{port}{the port used to run the ollama service. Default is 11374.}
}
\value{
A list that contains information about the ollama connection.
}
\description{
\code{get_ollama_connection} sets variables that are used to define the connection to an ollama server.
}
\details{
A simple function to set the connection details to an ollama server.
}
\examples{
#get_ollama_connection(ip_ad="127.0.0.1", port="11434")

}
