% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{get_ollama_embeddings}
\alias{get_ollama_embeddings}
\title{Get embeddings for a piece of context through an ollama server instance}
\usage{
get_ollama_embeddings(ollama_connection, model = "bge-large", input)
}
\arguments{
\item{ollama_connection}{a connection object that has the information on how to connect to the ollama server}

\item{model}{The model used to generates the embeddings. Defaults to "all-minilm". You can select any other model available for generating embeddings.}

\item{input}{A single character vector that contains the text to convert to embeddings}
}
\description{
\code{get_ollama_embeddings} Retrieves the embeddings for a string (can be a series of words, sentence, paragraph, or whole document) through ollama
}
\details{
Converts the ollama response for model info to a tibble
}
\examples{
conn <-get_ollama_connection()
embeddings <- get_ollama_embeddings(conn, input="Hi there, how are you doing?")
}
