% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{get_ollama_model_info}
\alias{get_ollama_model_info}
\title{Get information about one ollama model}
\usage{
get_ollama_model_info(ollama_connection, model)
}
\arguments{
\item{ollama_connection}{a connection object that has the information on how to connect to the Oollama server}

\item{model}{the name of the model that you want to get info from.}
}
\description{
\code{get_ollama_model_info} Gets information about one model available through ollama.
}
\details{
Gets information about one model available through ollama.
}
\examples{
#ollama_conn <- get_ollama_connection()
#get_ollama_model_info(ollama_connection= ollama_conn, model="llama3:8b-instruct-q5_0")

}
