% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{get_qdrant_connection}
\alias{get_qdrant_connection}
\title{Get Qdrant connection}
\usage{
get_qdrant_connection(
  endpoint = "http://localhost",
  port = 6333,
  api_key = NA_character_
)
}
\arguments{
\item{endpoint}{the URL pointing to the qdrant instance. Defaults to http://localhost}

\item{port}{the port to use to connect to the qdrant instance. Defaults to 6333}

\item{api_key}{optional for security. Defaults to NA. If given, it will use an API key to connect to the Qdrant.}
}
\description{
\code{get_qdrant_connection} establishes a connection with a qdrant instance.
'
}
\details{
Creates a connection object for qdrant.
If you are using the API key to connect to Qdrant, you need to specify it in the api_key parameter.
}
\examples{
conn <- get_qdrant_connection()
}
