% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{inspect_processing_skill}
\alias{inspect_processing_skill}
\title{Inspect a specific processing skill}
\usage{
inspect_processing_skill(processing_skill)
}
\arguments{
\item{processing_skill}{the name of the processing skill you want the model to use for a specific task}
}
\value{
a text vector containing the exact prompt used by that processing skill.
}
\description{
\code{inspect_processing_skill} lets you inspect how a specific processing skill is written.
}
\details{
This outputs the text contained in the processing skill. If the processing skill name you are looking for is not found, it will throw an error.
}
\examples{
inspect_processing_skill("break_down_tasks")

}
