% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{list_global_functions}
\alias{list_global_functions}
\title{List global functions}
\usage{
list_global_functions()
}
\description{
\code{list_global_functions} lists global functions available in the environment.
}
\details{
lists global functions available. This is useful to check if the functions used for tools calling are present or not.
}
\examples{
list_global_functions()

}
