% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{list_processing_skill_parameters}
\alias{list_processing_skill_parameters}
\title{List extra parameters for a given processing skill}
\usage{
list_processing_skill_parameters(processing_skill)
}
\arguments{
\item{processing_skill}{the name of the processing skill you want the model to use for a specific task}
}
\value{
a text vector containing the exact extra variables needed by that processing skill.
}
\description{
\code{list_processing_skill_parameters} returns the extra parameters required for a processing skill.
}
\details{
This outputs the variable names that need to be given as extra parameters to be able to use this processing skill.
If there is no extra variable, NA is returned.
}
\examples{
list_processing_skill_parameters("translate")

}
