% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{list_processing_skills}
\alias{list_processing_skills}
\title{list the processing skills}
\usage{
list_processing_skills()
}
\value{
a list of vectors describing the skills available currently in the package.
}
\description{
\code{list_processing_skills} lists the processing skills (i.e. prompts already defined) that you can use out of the box with this package.
}
\details{
This gives you a vector of all the skills available to use by default. This can then be used along with the set_processing_skill() function.
}
\examples{
list_processing_skills()

}
