% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{load_workflow}
\alias{load_workflow}
\title{Load workflow}
\usage{
load_workflow(filepath)
}
\arguments{
\item{filepath}{filepath to the JSON file that contains the parameters of the workflow, that you have created with save_workflow().}
}
\value{
the workflow after loading all of its settings and parameters.
}
\description{
\code{load_workflow} makes it possible to reload your workflow from a saved configuration file in JSON.
}
\details{
A function to reload your workflow from a specific JSON file that you have created with save_workflow_settings().
}
\examples{
my_workflow <- load_workflow(filepath="myworkflow.json")
}
