% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{parse_json_result}
\alias{parse_json_result}
\title{Parse JSON answer from the LLM}
\usage{
parse_json_result(json_string)
}
\arguments{
\item{json_string}{the JSON string you want to parse}
}
\description{
\code{parse_json_result} attempts to parse the JSON result from the LLM
}
\details{
This function will assume that the result from the LLM is provided in a JSON format
If the format is correct, it will parse the result as a R object.
You would typically expect such a format if you used the request_json_answer() function.
This is typically used in a pipe, after pull_final_answer().
}
\examples{
myflow_template <- ai_workflow() |> 
set_connector("ollama")  |> 
  set_model(model_name= "llama3.1:8b-instruct-q5_K_M") |> 
  set_n_predict(1000) |> 
  set_temperature(0.8) |> 
  request_json_answer()
  
myflow_template |> 
process_prompts("what is the usual color of the sky on Earth?") |>
pull_final_answer() |>
parse_json_result()

}
