% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{process_prompts}
\alias{process_prompts}
\title{Process Prompts starting from a workflow}
\usage{
process_prompts(workflow_obj, prompts)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the flow required}

\item{prompts}{A list that contains a potential mixture of text and image elements}
}
\description{
\code{process_prompts} is a way to process a vector of text prompts by starting from a workflow.
}
\details{
This function provides a way to process a text vector of prompts by starting from a workflow.
}
