% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{process_prompts_old}
\alias{process_prompts_old}
\title{Process Prompts starting from a workflow}
\usage{
process_prompts_old(workflow_obj, prompts_vector, images_vector = NA)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the flow required}

\item{prompts_vector}{A vector containing the prompts to be executed by the AI workflow}

\item{images_vector}{An optional vector (defaults to NA) containing the images to be send to the AI workflow}
}
\description{
\code{process_prompts_old} is a way to process a vector of prompts by starting from a workflow.
}
\details{
This function provides a way to process a vector of prompts by starting from a workflow.
}
