% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{process_text_prompts}
\alias{process_text_prompts}
\title{Process Text-only Prompts starting from a workflow}
\usage{
process_text_prompts(workflow_obj, text_prompts)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the flow required}

\item{text_prompts}{A vector containing the text prompts to be executed by the AI workflow}
}
\description{
\code{process_text_prompts} is a way to process a vector of text prompts by starting from a workflow.
}
\details{
This function provides a way to process a text vector of prompts by starting from a workflow.
}
