% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{pull_final_image_answer}
\alias{pull_final_image_answer}
\title{Pull Final Image Answer}
\usage{
pull_final_image_answer(workflow)
}
\arguments{
\item{workflow}{A workflow object containing all parameters describing the flow required}
}
\description{
\code{pull_final_image_answer} is a function to extract the final image resulting from a series of workflows
}
\details{
This function will extract the final resulting image from a series of workflows. It will ignore outputs of a different type.
}
