% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{qdrant_check_collection_existence}
\alias{qdrant_check_collection_existence}
\title{Qdrant: Check collection existence}
\usage{
qdrant_check_collection_existence(conn, collection_name = NA_character_)
}
\arguments{
\item{conn}{a connection object created by get_qdrant_connection()}

\item{collection_name}{the name of a collection for which you want to check the existence. Defaults to NA.}
}
\description{
\code{qdrant_check_collection_existence} establishes a connection with a qdrant instance and checks if a collection exists.
'
}
\details{
Confirms the existence of a qdrant collection.
Returns TRUE or FALSE
}
\examples{
conn <- get_qdrant_connection()
qdrant_check_collection_existence(conn, collection_name="hi_there")

}
