% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{qdrant_check_connection_validity}
\alias{qdrant_check_connection_validity}
\title{Qdrant: Check if the Connection is valid}
\usage{
qdrant_check_connection_validity(conn, silent = TRUE)
}
\arguments{
\item{conn}{a connection object created by get_qdrant_connection()}

\item{silent}{boolean, defaults to TRUE. If FALSE, prints a message when the connection is established.}
}
\description{
\code{qdrant_check_connection_validity} checks if the qdrant connection proposed is valid.
'
}
\details{
Does a simple test to confirm if the qdrant connection is working as expected.
}
\examples{
conn <- get_qdrant_connection()
qdrant_check_connection_validity(conn,silent=FALSE)
}
