% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{qdrant_create_new_collection}
\alias{qdrant_create_new_collection}
\title{Qdrant: Create new collection}
\usage{
qdrant_create_new_collection(
  conn,
  collection_name = NA_character_,
  vectors = list(size = 384, distance = "Cosine")
)
}
\arguments{
\item{conn}{a connection object created by get_qdrant_connection()}

\item{collection_name}{the name of the collection you want to create. Defaults to NA.}

\item{vectors}{a R list object that needs to contain 'size' and 'distance'. Defaults to size=384 and distance='Cosine'
This needs to be adapted to the embedding model you use. Different models have different resulting vector sizes.}
}
\description{
\code{qdrant_create_new_collection} establishes a connection with a qdrant instance and creates a new collection.
'
}
\details{
Creates a new collection for vector embeddings.
You can customize the size of embeddings needed and the type of distance for similarity search.
If a collection already exists with the same name it will throw an error.
}
\examples{
conn <- get_qdrant_connection()
qdrant_create_new_collection(conn, collection_name="story_of_my_life")
qdrant_delete_collection(conn, collection_name="story_of_my_life")

}
