% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{qdrant_delete_collection}
\alias{qdrant_delete_collection}
\title{Qdrant: Delete collection}
\usage{
qdrant_delete_collection(conn, collection_name = NA_character_)
}
\arguments{
\item{conn}{a connection object created by get_qdrant_connection()}

\item{collection_name}{the name of the collection you want to delete. Defaults to NA.}
}
\description{
\code{qdrant_delete_collection} establishes a connection with a qdrant instance and deletes an existing collection.
'
}
\details{
Deletes an existing collection on qdrant.
If the collection does not exists, it will throw an error.
If the deletion is successful it will display a confirmation message
}
\examples{
conn <- get_qdrant_connection()
qdrant_create_new_collection(conn, collection_name="story_of_alice")
qdrant_delete_collection(conn, collection_name="story_of_alice")

}
