% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{qdrant_delete_points}
\alias{qdrant_delete_points}
\title{Qdrant: Delete points (vectors)}
\usage{
qdrant_delete_points(conn, collection_name = NA_character_, ids = list())
}
\arguments{
\item{conn}{a connection object created by get_qdrant_connection()}

\item{collection_name}{the name of the collection you want to use. Defaults to NA.}

\item{ids}{a list of ids to delete. Defaults to an empty list.
The value of "ids" needs to be in this format: list(1,2,3) if you want to delete vectors 1,2 and 3.}
}
\description{
\code{qdrant_delete_points} deletes specific vectors (points)
'
}
\details{
This will delete specific vectors based on their ids.
}
\examples{
conn <- get_qdrant_connection()
qdrant_create_new_collection(conn, collection_name="test_db",
vectors=list(size=3,distance="Cosine"))
new_vectors <- list(points=list(
list(id=1, payload=list(text="hi there"),vector=list(0.1,0.5,0.6)),
list(id=2, payload=list(text="well well well"),vector=list(0.6,0.1,0.3))
))
qdrant_upsert_points(conn, points=new_vectors,collection_name="test_db")
qdrant_delete_points(conn, collection_name="test_db", ids=list(1,2))
qdrant_delete_collection(conn, "test_db")
}
