% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{qdrant_list_all_collections}
\alias{qdrant_list_all_collections}
\title{Qdrant: List all collections}
\usage{
qdrant_list_all_collections(conn)
}
\arguments{
\item{conn}{a connection object created by get_qdrant_connection()}
}
\description{
\code{qdrant_list_all_collections} establishes a connection with a qdrant instance and lists all collections.
'
}
\details{
Lists all the collections available on the qdrant instance.
Each collection can have different size of vector embeddings
}
\examples{
conn <- get_qdrant_connection()
qdrant_list_all_collections(conn)

}
