% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{qdrant_retrieve_point}
\alias{qdrant_retrieve_point}
\title{Qdrant: Retrieve a specific point (vector)}
\usage{
qdrant_retrieve_point(
  conn,
  collection_name = NA_character_,
  id = NA_character_
)
}
\arguments{
\item{conn}{a connection object created by get_qdrant_connection()}

\item{collection_name}{the name of the collection you want to use. Defaults to NA.}

\item{id}{the id of the point to retrieve. Defaults to NA.}
}
\description{
\code{qdrant_retrieve_point} retrieves a specific vector (point)
'
}
\details{
Retrieves the value of a specific vector/point, based on its id.
It will return a list with all the info available from qdrant.
}
\examples{
conn <- get_qdrant_connection()
qdrant_create_new_collection(conn, collection_name="test_db",
vectors=list(size=3,distance="Cosine"))
new_vectors <- list(points=list(list(id=1, payload=list(text="hi there"),
                      vector=list(0.1,0.5,0.6))))
qdrant_upsert_points(conn, points=new_vectors,collection_name="test_db",generate_id=FALSE)
qdrant_retrieve_point(conn, collection_name="test_db", id=1)
qdrant_delete_collection(conn, "test_db")

}
