% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdrant.R
\name{qdrant_search_points}
\alias{qdrant_search_points}
\title{Qdrant: Search points (vectors)}
\usage{
qdrant_search_points(
  conn,
  collection_name = NA_character_,
  vector = list(),
  limit = 5,
  with_payload = TRUE,
  with_vector = FALSE,
  score_threshold = NA_real_
)
}
\arguments{
\item{conn}{a connection object created by get_qdrant_connection()}

\item{collection_name}{the name of the collection you want to use. Defaults to NA.}

\item{vector}{a vector of embedding values that is used to initiate the search. Defaults to an empty list.
Note that this vector is simply a list containing numerical values for your embeddings.
If you use an embedding of size 384, this should be a simple list with 384 numerical values in it.}

\item{limit}{the maximum amount of vectors to return after the search. Defaults to 5.}

\item{with_payload}{if TRUE, will display the payload of the resulting vectors. Defaults to TRUE.}

\item{with_vector}{if TRUE, will display the full vector values in the results. Defaults to FALSE.}

\item{score_threshold}{a threshold for the minimum similarity score to return results. Defaults to NA.}
}
\description{
\code{qdrant_search_points} searches for similar vectors (points)
'
}
\details{
This will run a search for similar vectors on qdrant.
}
\examples{
conn <- get_qdrant_connection()
qdrant_create_new_collection(conn, collection_name="test_db",
vectors=list(size=3,distance="Cosine"))
new_vectors <- list(points=list(
list(id=1, payload=list(text="hi there"),vector=list(0.1,0.5,0.6)),
list(id=2, payload=list(text="well well well"),vector=list(0.6,0.1,0.3))
))
qdrant_upsert_points(conn, points=new_vectors, collection_name="test_db")
qdrant_search_points(conn, collection_name="test_db", vector=list(0.2,0.5,0.6))
qdrant_delete_collection(conn, "test_db")
}
