% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{request_json_answer}
\alias{request_json_answer}
\title{Request JSON answer from the LLM}
\usage{
request_json_answer(workflow_obj, json_object_format = list())
}
\arguments{
\item{workflow_obj}{the current workflow object that you want to build on}

\item{json_object_format}{the format required for the answer.}
}
\description{
\code{request_json_answer} request the current workflow to answer using a JSON format.
}
\details{
This function will request the LLM to answer using a JSON format. By default it will simply focus on a simple JSON format with just answer as a single object.
You can add a different JSON format as an argument.
}
\examples{
myflow_template <- ai_workflow() |> 
set_connector("ollama")  |> 
  set_model(model_name= "llama3.1:8b-instruct-q5_K_M") |> 
  set_n_predict(1000) |> 
  set_temperature(0.8) |> 
  request_json_answer()

}
