% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{save_workflow}
\alias{save_workflow}
\title{Save workflow}
\usage{
save_workflow(workflow_obj, filepath)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the workflow required}

\item{filepath}{filepath to a JSON file that will capture your workflow configuration.}
}
\value{
the workflow itself after saving it, so that it can be included inside a series of pipes.
}
\description{
\code{save_workflow} makes it possible to save your workflow settings for later re-use.
}
\details{
A function to save your workflow settings to a specific JSON file.
To reload the settings use the load_workflow() function.
}
\examples{
my_workflow <- ai_workflow() |> set_connector("ollama")  |>
 set_model(model_name= "llama3:8b-instruct-q5_K_S") |>
 set_n_predict(500) |>
 set_temperature(0.8) |> 
 set_default_missing_parameters_in_workflow() |> 
 set_system_prompt("You are a computer hardware and software specialist.") |>
 save_workflow(filepath="myworkflow.json")
}
