% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_audience}
\alias{set_audience}
\title{Define a specific audience you want the model to prepare an answer for}
\usage{
set_audience(workflow_obj, audience)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{audience}{a text description of the audience you want the LLM to write for.}
}
\value{
a workflow object with the new added audience parameter
}
\description{
\code{set_audience} lets you define for which audience the model should answer. This will tweak its answers accordingly to match their expectations.
}
\details{
This lets you modify a typical workflow by adding information about the expected audience.
}
\examples{
my_workflow <- ai_workflow() |> 
set_system_prompt(system_prompt="You are a helpful AI assistant. 
Answer to the best of your knowledge") |>
set_audience(audience="Marketing Professionals")

my_workflow <- ai_workflow() |> 
set_system_prompt(system_prompt="You are a helpful AI assistant. 
Answer to the best of your knowledge") |>
set_audience(audience="5 years old kids")

}
