% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_connector}
\alias{set_connector}
\title{Set the connector required to operate the workflow.}
\usage{
set_connector(workflow_obj, connector)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{connector}{the name of the connector type to use. It can be either ollama or llamacpp (to be taken care of in later versions).}
}
\description{
\code{set_connector} sets the connector type expected for the server calls. This can for example be an Ollama server or a Llama.cpp server.
}
\details{
Set the connector required to operate the workflow.
A connector is a server that will be reached by an API. It can either be a local or a remote server.
}
\examples{
my_workflow <- ai_workflow() |> 
set_connector(connector="ollama")
}
