% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_current_time_and_date_reference}
\alias{set_current_time_and_date_reference}
\title{Set the current time and date as addition reference}
\usage{
set_current_time_and_date_reference(workflow_obj)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the workflow required}
}
\value{
a workflow object with the new added current_timedate parameter
}
\description{
\code{set_current_time_and_date_reference} lets you set the current time and date as addition reference as part of the background for the LLM's answer.
}
\details{
Setting some additional time and date reference in the background info to be used by the LLM.
This can be very useful if you expect your LLM to answer about questions related to the future, present or past.
}
\examples{
my_workflow <- ai_workflow() |> 
set_system_prompt(system_prompt="You are a helpful AI assistant. 
Answer to the best of your knowledge") |>
set_current_time_and_date_reference()

}
