% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_custom_processing_skill}
\alias{set_custom_processing_skill}
\title{Set a custom processing skill (that you created) to give to the workflow.}
\usage{
set_custom_processing_skill(workflow_obj, filepath)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{filepath}{the file path where you have saved the processing skill. It needs to be a ".txt" file.}
}
\value{
the workflow object with the appropriate processing skill applied on the prompt vector(s)
}
\description{
\code{set_custom_processing_skill} sets a custom made processing skill to give to the workflow.
}
\details{
This sets a custom processing skill that you want to give to the workflow.
You can create a template of the skill file by using the create_custom_processing_skill() function.
}
\examples{
create_custom_processing_skill("write_poem.txt",
  system_prompt = "You are an AI specialized in writing poems like in the 18th century.",
  chat_prompt = "Write a poem based on the following text extract:", 
  final_guidance = "Only return the poem and nothing else. Do not add comments.")
my_workflow <- ai_workflow() |> 
  set_model(model_name="llama3:8b-instruct-q5_0") |> 
  set_custom_processing_skill("write_poem.txt")
  
}
