% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_default_missing_parameters_in_workflow}
\alias{set_default_missing_parameters_in_workflow}
\title{Set Defaults for missing workflow parameters}
\usage{
set_default_missing_parameters_in_workflow(workflow_obj, silent = F)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the flow required}

\item{silent}{ensure that the action happens without any warning sent to the standard output}
}
\value{
the workflow object along with the added parameters that were missing.
}
\description{
\code{set_default_missing_parameters_in_workflow} ensure all parameters are present by filling the gaps
}
\details{
A function to add default values for missing parameters in the user-defined workflow.
}
\examples{
# my_workflow <- ai_workflow() |> set_default_missing_parameters_in_workflow()
}
