% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_ip_addr}
\alias{set_ip_addr}
\title{Set the IP Address required to connect to an API server.}
\usage{
set_ip_addr(workflow_obj, ip_addr)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{ip_addr}{IP address of the server.}
}
\description{
\code{set_ip_addr} sets the IP Address related to the API server.
}
\details{
Set the IP Address related to the API server.
}
\examples{
wflow <- ai_workflow() |> set_connector("ollama") |> 
set_ip_addr(ip_addr="127.0.0.1")
wflow <- ai_workflow() |> set_connector("ollama") |> 
set_ip_addr(ip_addr="192.168.1.56")

}
