% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_mode}
\alias{set_mode}
\title{Set the mode of the model used by the workflow.}
\usage{
set_mode(workflow_obj, mode)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{mode}{the mode to be used by the model in the workflow - it needs to be either 'chat', 'completion' or 'embeddings'.}
}
\value{
the workflow object with the mode to be used applied as a new parameter.
}
\description{
\code{set_mode} sets the mode of the model in the workflow. This usually means either 'chat', 'completion' or 'embeddings'.
}
\details{
This sets the mode of the model in the workflow. This usually means either 'chat','completion' or 'embeddings'.
}
\examples{
my_workflow <- ai_workflow() |> 
set_model(model_name="llama3:8b-instruct-q5_0") |> 
set_mode("chat")
}
