% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_model}
\alias{set_model}
\title{Set the LLM model to be used by the workflow}
\usage{
set_model(workflow_obj, model_name)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{model_name}{the name of the model to use in the workflow}
}
\description{
\code{set_model} sets the model to be used by a workflow object.
}
\details{
A simple function to set the model to be used. Note that this model needs to be available at the instance you connect to.
}
\examples{
my_workflow <- ai_workflow() |> 
set_model(model_name="llama3:8b-instruct-q5_0")

}
