% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_n_predict}
\alias{set_n_predict}
\title{Set the number of tokens to be predicted (maximum) by the flow.}
\usage{
set_n_predict(workflow_obj, n_predict)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{n_predict}{the number of tokens to be predicted (maximum) by the flow.}
}
\description{
\code{set_n_predict} sets the maximum number of tokens to be predicted by the flow.
}
\details{
This sets the maximum number of tokens to be predicted by the flow. Note that this does not mean that the LLM will constrain its answer to that number,
If the planned answer exceeds the n_predict value, the answer will simply stop at that point.
}
\examples{
wflow <- ai_workflow() |> set_connector("ollama") |> 
set_n_predict(n_predict=500)
}
