% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_port}
\alias{set_port}
\title{Set the port required to connect to the API server.}
\usage{
set_port(workflow_obj, port)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{port}{the port used to access the API server.}
}
\description{
\code{set_port} sets the port related to the API server.
}
\details{
Set the IP Address related to the API server.
}
\examples{
wflow <- ai_workflow() |> set_connector("ollama") |> 
set_ip_addr(ip_addr="127.0.0.1") |> set_port(port="11434")
}
