% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_presence_penalty}
\alias{set_presence_penalty}
\title{Set the presence penalty of the model used by the flow.}
\usage{
set_presence_penalty(workflow_obj, presence_penalty)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{presence_penalty}{the presence penalty value used by the model. A value closer to 1 encourages the model to generate more novel and diverse text. A lower value, closer to 0, encourages cliche phrases.}
}
\description{
\code{set_presence_penalty} sets the presence penalty related to the model in the workflow.
}
\details{
Set the presence_penalty to be used by the model in the workflow.
}
\examples{
my_workflow <- ai_workflow() |> 
set_model(model_name="llama3:8b-instruct-q5_0") |> 
set_presence_penalty(0.9)

}
