% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_processing_skill}
\alias{set_processing_skill}
\title{Set the processing skill that you want to give the workflow.}
\usage{
set_processing_skill(workflow_obj, processing_skill, ...)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{processing_skill}{string - the processing skill you want to give to the workflow.}

\item{...}{additional parameters that will be treated if they are supported by the processing skill}
}
\value{
the workflow object with the appropriate processing skill applied on the prompt vector(s)
}
\description{
\code{set_processing_skill} sets the processing skill to give to the workflow.
}
\details{
This sets the processing skill that you want to give the workflow. You can list up the processing skills available by default with list_processing_skills().
}
\examples{
my_workflow <- ai_workflow() |> set_model(model_name="llama3:8b-instruct-q5_0") |> 
set_processing_skill("fix_copy")
}
