% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_repeat_penalty}
\alias{set_repeat_penalty}
\title{Set the repeat penalty of the model used by the flow.}
\usage{
set_repeat_penalty(workflow_obj, repeat_penalty)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{repeat_penalty}{the repeat penalty value used by the model. A value of 1 means there is no penalty. A value higher than 1 means there is increased penalty on repetition of tokens.}
}
\description{
\code{set_repeat_penalty} sets the repeat penalty related to the model in the workflow.
}
\details{
Set the repeat_penalty to be used by the model in the workflow.
}
\examples{
my_workflow <- ai_workflow() |> 
set_model(model_name="llama3:8b-instruct-q5_0") |> 
set_repeat_penalty(1.3)

}
