% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_system_prompt}
\alias{set_system_prompt}
\title{Set the system prompt to be used by the model.}
\usage{
set_system_prompt(workflow_obj, system_prompt)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{system_prompt}{the mode to be used by the model in the workflow - it needs to be either 'chat' or 'completion'.}
}
\value{
the workflow object with the system prompt specified as new parameter.
}
\description{
\code{set_system_prompt} sets the system prompt to be used by the model.
}
\details{
This sets the system prompt of the model in the workflow. This is where you can give additional guidance or a personality to change the way your model will answer prompts.
}
\examples{
my_workflow <- ai_workflow() |> 
set_system_prompt(system_prompt="You are a helpful AI assistant. 
Answer to the best of your knowledge")
my_workflow <- ai_workflow() |> 
set_system_prompt(system_prompt="You are a helpful AI assistant
 and an expert in financial analysis.")
}
