% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{split_text_as_sentences}
\alias{split_text_as_sentences}
\title{Split text into sentences}
\usage{
split_text_as_sentences(text)
}
\arguments{
\item{text}{a single piece of text to break down.}
}
\description{
\code{split_text_as_sentences} Splits text into sentences-based chunks.
}
\details{
Splits text into sentences-based chunks, leveraging the tokenizers library.
}
\examples{
# c("Hi! How are you? Do you want to go for a walk?") |> split_text_as_sentences()

}
