% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{switch_to_workflow}
\alias{switch_to_workflow}
\title{Switch to workflow}
\usage{
switch_to_workflow(workflow_obj, new_workflow)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the flow required}

\item{new_workflow}{A workflow object to execute after the last one}
}
\description{
\code{switch_to_workflow} is a function that makes it possible to chain several workflows once after the other using pipes
}
\details{
This function will send the output from the previous workflow to the next one.
}
