% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamacpp.R
\name{test_llamacpp_connection}
\alias{test_llamacpp_connection}
\title{Confirm connection to a Llama.cpp server is working}
\usage{
test_llamacpp_connection(ip_ad = "127.0.0.1", port = "8080")
}
\arguments{
\item{ip_ad}{the IP address of the server running Llama.cpp. Default is localhost 127.0.0.1.}

\item{port}{the port used to run the Llama.cpp service. Default is 8080.}
}
\description{
\code{test_llamacpp_connection} tests a connection to a llama.cpp server.
}
\details{
A simple function to test the connection to a llama.cpp server.
}
