% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{test_ollama_connection}
\alias{test_ollama_connection}
\title{Confirm connection to ollama is working}
\usage{
test_ollama_connection(ip_ad = "127.0.0.1", port = "11434")
}
\arguments{
\item{ip_ad}{the IP address of the server running ollama. Default is localhost 127.0.0.1.}

\item{port}{the port used to run the ollama service. Default is 11374.}
}
\value{
Returns TRUE if the connection works, or an error object it does not.
}
\description{
\code{test_ollama_connection} says hello and uses the name of the person(s) as an argument. y
}
\details{
A simple function to test the connect to ollama
}
\examples{
#test_ollama_connection(ip_ad="127.0.0.1", port="11434")

}
