% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{unload_models_automatically}
\alias{unload_models_automatically}
\title{Unload models automatically inside a workflow}
\usage{
unload_models_automatically(workflow_obj)
}
\arguments{
\item{workflow_obj}{A workflow object containing all parameters describing the flow required}
}
\description{
\code{unload_models_automatically} takes care of freeing memory after each model has been used, in order to avoid out of memory problems.
}
\details{
This function provides a parameter to free memory after each model has been used, in order to avoid out of memory problems.
}
