% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{pull_final_answer}
\alias{pull_final_answer}
\title{Pull Final Answer}
\usage{
pull_final_answer(workflow)
}
\arguments{
\item{workflow}{A workflow object containing all parameters describing the flow required}
}
\description{
\code{pull_final_answer} is a function to extract the final answer from a series of workflows
}
\details{
This function will extract the final text result that's usually coming from a series of workflows.
}
