% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{set_seed}
\alias{set_seed}
\title{Set the seed of the model used by the workflow.}
\usage{
set_seed(workflow_obj, seed)
}
\arguments{
\item{workflow_obj}{an ai_workflow object created by ai_workflow() in the first place.}

\item{seed}{the seed to be used by the model in the workflow, if you intend to fix it. The same seed will give the same answer. It is interesting to fix it if you want to compare the effect of other parameters.}
}
\value{
the workflow object with the seed to be applied.
}
\description{
\code{set_seed} sets the seed for the model.
}
\details{
Set the seed to be used by the model in the workflow.
}
\examples{
my_workflow <- ai_workflow() |> 
set_model(model_name="llama3:8b-instruct-q5_0") |> 
set_seed(12312312312)
}
